feat: core refactored from ak to core and cli modules
