+++ ak-py/src/ak/__init__.py
@@ -18,4 +18,4 @@ try:
 except importlib.metadata.PackageNotFoundError:
     __version__ = "0.1.0"
 
-from .ak import *
+from .core import *
