+++ ak-py/src/ak/cli/cli.py
@@ -1,8 +1,10 @@
 import asyncio
 import readline  # Enables line editing and history features for input() in the CLI
-import uuid
 
-from .ak import Runtime, Session
+from uuid import uuid4
+
+from ..core import Runtime, Session
+
 
 class CLI:
     """
@@ -44,7 +46,7 @@ class CLI:
 
     def new(self):
         if self._agent:
-            self._session = Session(str(uuid.uuid4()))
+            self._session = Session(str(uuid4()))
             print(f"Starting new session: {self._session.id}")
         else:
             print("No agent selected. Please select an agent using !select <agent_name>.")
@@ -121,6 +123,7 @@ class CLI:
             asyncio.run(cli.run())
         except asyncio.CancelledError:
             print()
+
 
 if __name__ == "__main__":
     CLI.main()
