+++ ak-py/src/ak/core/base.py
@@ -1,8 +1,6 @@
-import logging
-
 from abc import abstractmethod
-from importlib import import_module
 from typing import Any
+
 
 class Session:
     """
@@ -50,6 +48,7 @@ class Session:
         self._data[key] = value
         return value
 
+
 class Runner:
     """
     Runner is the base class for all agent runners.
@@ -85,6 +84,7 @@ class Runner:
         """
         pass
 
+
 class Agent:
     """
     Agent is the base class for all agents.
@@ -117,78 +117,3 @@ class Agent:
         Returns the runner associated with the agent.
         """
         return self._runner
-
-class Module:
-    """
-    Module is the base class for all agent modules.
-    
-    An agent module is a Python module containing a set of agents built using a supported agent
-    framework. Agent Kernel provides an implementation of the Module class for each supported agent
-    framework, allowing the agents to be registered and managed in a framework agnostic manner.
-    """
-
-    def __init__(self, agents: list[Agent]):
-        """
-        Initializes a Module instance.
-        :param agents: List of agents in the module.
-        """
-        self._agents = agents
-        for agent in agents:
-            Runtime.register(agent)
-
-    @property
-    def agents(self) -> list[Agent]:
-        """
-        Returns the list of agents in the module.
-        """
-        return self._agents
-
-class Runtime:
-    """
-    Runtime class provides the environment for hosting and running agents.
-    """
-
-    _log = logging.getLogger("ak.runtime")
-    _agents = {}
-
-    @staticmethod
-    def load(module):
-        """
-        Loads an agent module dynamically.
-        :param module: Name of the module to load.
-        :return: The loaded module.
-        """
-        Runtime._log.debug(f"Loading module '{module}'")
-        return import_module(module)
-
-    @staticmethod
-    def register(agent: Agent):
-        """
-        Registers an agent in the runtime.
-        :param agent: The agent to register.
-        """
-        if not Runtime._agents.get(agent.name):
-            Runtime._log.debug(f"Registering agent '{agent.name}'")
-            Runtime._agents[agent.name] = agent
-        else:
-            Runtime._log.warning(f"Agent with name '{agent.name}' is already registered.")
-
-    @staticmethod
-    def agents() -> dict[str, Agent]:
-        """
-        Returns the list of registered agents.
-        :return: List of agents.
-        """
-        return Runtime._agents
-
-    @staticmethod
-    async def run(agent: Agent, session: Session, prompt: Any) -> Any:
-        """
-        Runs the specified agent with the given prompt.
-        :param agent: The agent to run.
-        :param session: The session to use for the agent.
-        :param prompt: The prompt to provide to the agent.
-        :return: The result of the agent's execution.
-        """
-        Runtime._log.debug(f"Running agent '{agent.name}' with prompt: {prompt}")
-        return await agent.runner.run(agent, session, prompt)
