+++ ak-py/src/ak/crewai/__init__.py
@@ -13,4 +13,4 @@ try:
 except importlib.metadata.PackageNotFoundError:
     __version__ = "0.1.0"
 
-from .akcrewai import CrewAIModule as AgentModule
+from .crewai import CrewAIModule as AgentModule
