+++ ak-py/src/ak/langgraph/__init__.py
@@ -1,7 +1,7 @@
 """
-Agent Kernel support for LangGraph Agents SDK.
+Agent Kernel support for LangGraph.
 
-This package contains Agent Kernel support for agents built with LangGraph Agents SDK.
+This package contains Agent Kernel support for agents built with LangGraph.
 It provides the necessary classes and methods to integrate LangGraph agents into the Agent Kernel
 framework, allowing for seamless interaction and execution of LangGraph based agents.
 """
@@ -13,4 +13,4 @@ try:
 except importlib.metadata.PackageNotFoundError:
     __version__ = "0.1.0"
 
-from .aklanggraph import LangGraphModule as AgentModule, LangGraphRunner as AgentRunner, LangGraphAgent as Agent
+from .langgraph import LangGraphModule as AgentModule
