+++ ak-py/src/ak/openai/__init__.py
@@ -13,4 +13,4 @@ try:
 except importlib.metadata.PackageNotFoundError:
     __version__ = "0.1.0"
 
-from .akopenai import OpenAIModule as AgentModule
+from .openai import OpenAIModule as AgentModule
