+++ examples/aws-serverless/openai/lambda.py
@@ -1,4 +1,5 @@
 from agents import Agent
+
 from ak.aws import Lambda
 from ak.openai import AgentModule
