+++ examples/cli/crewai/pyproject.toml
@@ -6,7 +6,7 @@ description = "Demo of CrewAI based agents running in Agent Kernel with a CLI"
 readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
-    "ak[crewai]>=0.1.0",
+    "ak[cli,crewai]>=0.1.0",
 ]
 
 [tool.uv]
