+++ examples/cli/langgraph/custom_agent.py
@@ -1,4 +1,5 @@
 from typing import Any, Sequence, TypedDict, Annotated, Callable, List
+
 from langchain_core.messages import BaseMessage, SystemMessage
 from langchain_core.tools import StructuredTool
 from langgraph.graph import StateGraph, END, add_messages
