+++ examples/cli/langgraph/demo.py
@@ -1,8 +1,9 @@
-from ak import CLI
-from ak.langgraph import AgentModule
 from langchain_openai import ChatOpenAI
 from langgraph.prebuilt import create_react_agent
 from langgraph_supervisor import create_supervisor
+
+from ak.cli import CLI
+from ak.langgraph import AgentModule
 
 from custom_agent import CustomAgent
