+++ examples/cli/langgraph/pyproject.toml
@@ -6,7 +6,7 @@ description = "Demo of LangGraph based agents running in Agent Kernel with a CLI
 readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
-    "ak[langgraph]>=0.1.0",
+    "ak[cli,langgraph]>=0.1.0",
     "langchain-litellm>=0.2.2",
     "langchain-openai>=0.3.28",
     "langgraph-supervisor>=0.0.29",
