+++ examples/cli/openai/pyproject.toml
@@ -6,7 +6,7 @@ description = "Demo of OpenAI Agents SDK based agents running in Agent Kernel wi
 readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
-    "ak[openai]>=0.1.0",
+    "ak[cli,openai]>=0.1.0",
 ]
 
 [tool.uv]
